HOUSE RENT PREDICTION/requirements.txt
gunicorn
flask
flask-sqlalchemy
flask-login
flask-wtf
flask-bcrypt
folium
python-dotenv
flask-mail
flask-migrate
flask-cors
WTForms-Components
pillow
scikit-learn==1.7.1
pandas
numpy
xgboost
matplotlib
seaborn
flask-uploads
PyJWT
flask_dance
dataframe-image
